#4: Move application startup/teardown code from Application into separate classes
